pyproject.toml
[project]
name = "mcpvenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastembed>=0.6.1",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.8",
    "mcp[cli]>=1.7.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
]
